fixe warn
git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@4926 534d5172-be3c-0410-9a3e-ccaa6372a113

+++ src/utils/probe/browserthread.cpp
@@ -176,7 +176,7 @@ BrowserThread::filterRegistry()
 void 
 BrowserThread::loadPlatform()
 {
-    if ( pick_ >= registryData_.platforms.size() ) {
+    if ( pick_ >= (int)registryData_.platforms.size() ) {
         events_->add( Fault );
         return;
     }
